Integrated Webscraper and search results with html
Now when, the user searches enters a search query, the website takes it and runs it through the scraper. The scraper searches google for the exact input by the user (this is all happening in the back-end with no mention of google to the user). It then takes all the links available on google's search results, strips them down to just the hyper links and sends those hyperlinks back to the program.

Then, the program iteratively takes each individual link and slices away any suffix and prefixes that googled added as a security measure (they do this to break links so you can't use them). All the (fully-functioning) hyper links are then added to a dictionary and printed on to the next page with indexing.